Decouple Jackson from XContent + parametric tests
